MacOS bug fixes

+++ src/functions.h
@@ -39,7 +39,7 @@ results* get_dir(char *pwd);
 results* reorder_ob(results* ob, char *order);
 void display_dir(char *pwd, results* ob, int topfileref, int selected);
 void set_history(char *pwd, int topfileref, int selected);
-long GetAvailableSpace(const char* path);
+size_t GetAvailableSpace(const char* path);
 long GetUsedSpace(const char* path);
 int SendToPager(const char* object);
 int SendToEditor(const char* object);
